[NU-2189] ScenarioTestingApiHttpServiceSpec simpler: - extracted separate FragmentScenarioTestingApiHttpServiceSpec class instead of running fragments test for event-generator (in 4 variants) and for GenericSourceWithCustomTestingSupport - SchemalessKafkaJsonTypeTests: removed WithSchemalessAdHocTestsLogic abstraction layer which was used only by SchemalessKafkaJsonTypeTests
