+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/testing/ScenarioTestingApiHttpServiceSpec.scala
@@ -9,13 +9,11 @@ import org.apache.pekko.http.scaladsl.model.StatusCodes
 import org.scalatest.Assertion
 import org.scalatest.freespec.AnyFreeSpecLike
 import org.scalatest.matchers.should.Matchers
-import pl.touk.nussknacker.engine.api.definition.FixedExpressionValue
 import pl.touk.nussknacker.engine.api.graph.ScenarioGraph
-import pl.touk.nussknacker.engine.api.parameter.{ParameterName, ValueInputWithFixedValuesProvided}
+import pl.touk.nussknacker.engine.api.parameter.ParameterName
 import pl.touk.nussknacker.engine.api.typed.typing.TypingResult
 import pl.touk.nussknacker.engine.build.ScenarioBuilder
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
-import pl.touk.nussknacker.engine.graph.node.FragmentInputDefinition.{FragmentClazzRef, FragmentParameter}
 import pl.touk.nussknacker.test.{
   NuRestAssureMatchers,
   PatientScalaFutures,
@@ -60,36 +58,6 @@ trait ScenarioTestingApiHttpServiceSpec
   protected def expectedTestDataJson: String
 
   protected val inputValueNotMatchingExpectedType = "false"
-
-  private val fragmentFixedParameter = FragmentParameter(
-    ParameterName("paramFixedString"),
-    FragmentClazzRef[java.lang.String],
-    initialValue = Some(FixedExpressionValue("'uno'", "uno")),
-    hintText = None,
-    valueEditor = Some(
-      ValueInputWithFixedValuesProvided(
-        fixedValuesList = List(
-          FixedExpressionValue("'uno'", "uno"),
-          FixedExpressionValue("'due'", "due"),
-        ),
-        allowOtherValue = false
-      )
-    ),
-    valueCompileTimeValidation = None
-  )
-
-  private val fragmentRawStringParameter = FragmentParameter(
-    ParameterName("paramRawString"),
-    FragmentClazzRef[java.lang.String],
-    initialValue = None,
-    hintText = None,
-    valueEditor = None,
-    valueCompileTimeValidation = None
-  )
-
-  private def exampleFragment(parameter: FragmentParameter) = ScenarioBuilder
-    .fragmentWithRawParameters("fragment", parameter)
-    .fragmentOutput("fragmentEnd", "output", "out" -> "'hola'".spel)
 
   "The endpoint for capabilities should" - {
     "return valid capabilities for scenario with all capabilities" in {
@@ -215,163 +183,6 @@ trait ScenarioTestingApiHttpServiceSpec
              |]
              |""".stripMargin
           )
-        )
-    }
-    "generate parameters for fragment with fixed list parameter" in {
-      val fragment = exampleFragment(fragmentFixedParameter)
-      val expectedTestParameters =
-        s"""[
-             |    {
-             |        "sourceId": "fragment",
-             |        "parameters": [
-             |            {
-             |                "name": "paramFixedString",
-             |                "typ": {
-             |                    "display": "String",
-             |                    "type": "TypedClass",
-             |                    "refClazzName": "java.lang.String",
-             |                    "params": [
-             |
-             |                    ]
-             |                },
-             |                "editors": [{
-             |                    "possibleValues": [
-             |                        {
-             |                            "expression": "",
-             |                            "label": ""
-             |                        },
-             |                        {
-             |                            "expression": "'uno'",
-             |                            "label": "uno"
-             |                        },
-             |                        {
-             |                            "expression": "'due'",
-             |                            "label": "due"
-             |                        }
-             |                    ],
-             |                    "type": "FixedValuesParameterEditor"
-             |                }],
-             |                "defaultValue": {
-             |                    "language": "spel",
-             |                    "expression": "'uno'"
-             |                },
-             |                "additionalVariables": {
-             |
-             |                },
-             |                "variablesToHide": [
-             |
-             |                ],
-             |                "branchParam": false,
-             |                "hintText": null,
-             |                "label": "paramFixedString",
-             |                "requiredParam": false,
-             |                "category": "Standard",
-             |                "changesCanReloadParameters": false,
-             |                "nonImportantForExecution": false
-             |            }
-             |        ]
-             |    }
-             |]
-             |""".stripMargin
-      given()
-        .applicationState {
-          createSavedScenario(fragment)
-        }
-        .when()
-        .basicAuthAllPermUser()
-        .jsonBody(canonicalGraphStr(fragment))
-        .post(s"$nuDesignerHttpAddress/api/scenarioTesting/${fragment.name}/capabilities")
-        .Then()
-        .statusCode(200)
-        .equalsJsonBody(
-          s"""{
-             |    "testWithParameters": {
-             |      "status": "AVAILABLE",
-             |      "sourceParameters": $expectedTestParameters
-             |    },
-             |    "testWithGeneratedData": {
-             |      "status": "NOT_AVAILABLE",
-             |      "reason":"NOT_SUPPORTED_BY_SOURCES"
-             |    },
-             |    "testWithLiveData": {
-             |      "status": "NOT_AVAILABLE",
-             |      "reason":"NOT_SUPPORTED_BY_SOURCES"
-             |    }
-             |}""".stripMargin
-        )
-    }
-    "Generate parameters with simplified (single) editor for fragment with raw string parameter" in {
-      val fragment = exampleFragment(fragmentRawStringParameter)
-      val expectedTestParameters =
-        s"""[
-           |    {
-           |        "sourceId": "fragment",
-           |        "parameters": [
-           |            {
-           |                "name": "paramRawString",
-           |                "typ": {
-           |                    "display": "String",
-           |                    "type": "TypedClass",
-           |                    "refClazzName": "java.lang.String",
-           |                    "params": [
-           |
-           |                    ]
-           |                },
-           |                "editors": [
-           |                    {
-           |                        "type": "SpelTemplateParameterEditor"
-           |                    },
-           |                    {
-           |                        "type": "SpelParameterEditor"
-           |                    }
-           |                ],
-           |                "defaultValue": {
-           |                    "language": "spelTemplate",
-           |                    "expression": ""
-           |                },
-           |                "additionalVariables": {
-           |
-           |                },
-           |                "variablesToHide": [
-           |
-           |                ],
-           |                "branchParam": false,
-           |                "hintText": null,
-           |                "label": "paramRawString",
-           |                "requiredParam": false,
-           |                "category": "Standard",
-           |                "changesCanReloadParameters": false,
-           |                "nonImportantForExecution": false
-           |            }
-           |        ]
-           |    }
-           |]
-           |""".stripMargin
-      given()
-        .applicationState {
-          createSavedScenario(fragment)
-        }
-        .when()
-        .basicAuthAllPermUser()
-        .jsonBody(canonicalGraphStr(fragment))
-        .post(s"$nuDesignerHttpAddress/api/scenarioTesting/${fragment.name}/capabilities")
-        .Then()
-        .statusCode(200)
-        .equalsJsonBody(
-          s"""{
-             |    "testWithParameters": {
-             |      "status": "AVAILABLE",
-             |      "sourceParameters": $expectedTestParameters
-             |    },
-             |    "testWithGeneratedData": {
-             |      "status": "NOT_AVAILABLE",
-             |      "reason":"NOT_SUPPORTED_BY_SOURCES"
-             |    },
-             |    "testWithLiveData": {
-             |      "status": "NOT_AVAILABLE",
-             |      "reason":"NOT_SUPPORTED_BY_SOURCES"
-             |    }
-             |}""".stripMargin
         )
     }
     "return error if scenario does not exists" in {
@@ -540,9 +351,6 @@ trait ScenarioTestingApiHttpServiceSpec
        |  "scenarioGraph": $scenarioGraphStr,
        |  "numberOfSamples": $numberOfSamples
        |}""".stripMargin
-
-  private def canonicalGraphStr(canonical: CanonicalProcess) =
-    Encoder[ScenarioGraph].apply(canonical.toScenarioGraph).toString()
 
   protected def exampleScenarioInputVariableType: TypingResult
